Function    fspan(start [float], finish [float], spacing [int]) changed to:             fspan(start [numeric], finish [numeric], spacing [int])
to accomodate promotion of arguments to type double.  When a user has data of
type double, precision is lost if it is demoted to type float.  If either of
the arguments "start" or "finish" are of type double, computation is done with
type double.  Any other numeric type is promoted to type float for computation.

This change also fixes the problem of integer variables being given as the
"start" and "finish" arguments generating a somewhat cryptic coercion message.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6993 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddBuiltIns.c
@@ -1,6 +1,6 @@
 
 /*
- *      $Id: AddBuiltIns.c,v 1.68 2003-05-31 00:49:40 dbrown Exp $
+ *      $Id: AddBuiltIns.c,v 1.69 2003-06-06 20:59:55 grubin Exp $
  */
 /************************************************************************
 *                                                                       *
@@ -1360,10 +1360,10 @@ void _NclAddBuiltIns
 	nargs = 0;
 	args = NewArgs(3);
 	dimsizes[0] = 1;
-	SetArgTemplate(args,nargs,"float",1,dimsizes); nargs++;
-	SetArgTemplate(args,nargs,"float",1,dimsizes); nargs++;
-	SetArgTemplate(args,nargs,"integer",1,dimsizes); nargs++;
-	NclRegisterFunc( _Nclfspan,args,"fspan",nargs);
+	SetArgTemplate(args,nargs, "numeric", 1, dimsizes);  nargs++;
+	SetArgTemplate(args,nargs, "numeric", 1, dimsizes);  nargs++;
+	SetArgTemplate(args,nargs, "integer", 1, dimsizes);  nargs++;
+	NclRegisterFunc( _Nclfspan, args, "fspan", nargs);
 
 	nargs = 0;
 	args = NewArgs(3);
